feat: allow for running any number of route handlers
