+++ crates/nest/src/lib.rs
@@ -1,16 +1,23 @@
 pub mod config;
 
 pub mod http;
-// lib.rs or http/mod.rs
+
+use anyhow::{Context, Result};
+use sqlx::sqlite::SqlitePoolOptions;
+use tracing_subscriber::{
+    filter::EnvFilter, fmt::format::JsonFields, util::SubscriberInitExt, Layer
+};
+
 use axum::Router;
 use sqlx::SqlitePool;
 use std::sync::Arc;
 use std::collections::HashMap;
+use std::future::Future;
+use std::pin::Pin;
 use tokio::net::TcpListener;
 use tower::ServiceBuilder;
 use tower_http::add_extension::AddExtensionLayer;
 use tower_http::trace::TraceLayer;
-use anyhow::Result;
 use crate::config::Config;
 
 pub trait MultiDbHandler: Send + Sync + 'static {
@@ -29,6 +36,14 @@ pub struct ApiBuilder {
     handlers: Vec<Box<dyn MultiDbHandler>>,
 }
 
+pub trait Run {
+    type Future: Future<Output = Result<()>> + Send + 'static;
+
+    fn run<I>(config: Config, handlers: I) -> Self::Future
+    where
+        I: IntoIterator<Item = Box<dyn MultiDbHandler + Send>> + Send + 'static;
+}
+
 impl ApiBuilder {
     pub fn new(config: Config) -> Self {
         Self {
@@ -43,8 +58,8 @@ impl ApiBuilder {
         self
     }
 
-    pub fn add_handler<H: MultiDbHandler + 'static>(&mut self, handler: H) -> &mut Self {
-        self.handlers.push(Box::new(handler));
+    pub fn add_handler(&mut self, handler: Box<dyn MultiDbHandler>) -> &mut Self {
+        self.handlers.push(handler);
         self
     }
 
@@ -74,4 +89,71 @@ impl ApiBuilder {
 
         Ok(())
     }
+}
+
+impl Run for ApiBuilder {
+    type Future = Pin<Box<dyn Future<Output = Result<()>> + Send + 'static>>;
+
+    // Now the run method accepts handlers
+    fn run<I>(config: Config, handlers: I) -> Self::Future
+    where
+        I: IntoIterator<Item = Box<dyn MultiDbHandler + Send>> + Send + 'static,
+    {
+        Box::pin(async move {
+            let mut api_builder = Self::new(config.clone());
+
+            for (i, db_url) in config.database_urls.split(',').enumerate() {
+                let db_name = format!("db_{}", i + 1);
+                let db_pool = SqlitePoolOptions::new()
+                    .max_connections(5)
+                    .connect(db_url)
+                    .await
+                    .context(format!("could not connect to {}", db_url))?;
+
+                sqlx::migrate!().run(&db_pool).await?;
+                api_builder.add_db(db_name, db_pool);
+            }
+
+            // Add the dynamic handlers passed into the function
+            for handler in handlers {
+                api_builder.add_handler(handler);
+            }
+
+            api_builder.serve().await?; // Serve the API
+
+            Ok(())
+        })
+    }
+}
+
+pub fn init_tracing(config: &Config) -> Option<()> {
+    use std::str::FromStr;
+    use tracing_subscriber::layer::SubscriberExt;
+
+    let filter = EnvFilter::from_str(config.rust_log.as_deref()?).unwrap();
+
+    tracing_subscriber::registry()
+        .with(
+            if config.log_json {
+                Box::new(
+                    tracing_subscriber::fmt::layer()
+                    .fmt_fields(JsonFields::default())
+                    .event_format(
+                        tracing_subscriber::fmt::format()
+                            .json()
+                            .flatten_event(true)
+                            .with_span_list(false),
+                    )
+                    .with_filter(filter),
+                ) as Box<dyn Layer<_> + Send + Sync>
+            } else {
+                Box::new(
+                    tracing_subscriber::fmt::layer()
+                        .event_format(tracing_subscriber::fmt::format().pretty())
+                        .with_filter(filter),
+                )
+            }
+        )
+        .init();
+    None
 }
